+++ src/FunFair.BuildCheck/Setup.cs
@@ -99,6 +99,7 @@ namespace FunFair.BuildCheck
             AddProjectCheck<ImportCommonProps>(services);
             AddProjectCheck<PublishableExesMustHaveRuntimeIdentifiers>(services);
             AddProjectCheck<ShouldNotRemoveFromCompilation>(services);
+            AddProjectCheck<IsTransformWebConfigDisabledShouldBeTrueForWebLibraryProjects>(services);
         }
 
         private static bool IsEarlierThanDotNet5()